Correct a few casts.

+++ src/GAME5.c
@@ -13924,7 +13924,7 @@ int sub_578CD0()
         if (sub_4CB230(v3, v4))
         {
             sub_4B0300(v4);
-            sub_4B0640((int)sub_578C60);
+            sub_4B0640(sub_578C60);
             result = sub_4B0340(1);
         }
         else
